+++ src/main/java/com/chrisrm/ideaddlc/MTAnalytics.java
@@ -70,9 +70,6 @@ public final class MTAnalytics {
   public static final String INDICATOR_STYLE = "IndicatorStyle";
   public static final String SELECT_THEME = "SelectTheme";
   public static final String HELP = "Help";
-  public static final String COMPACT_TABLES = "CompactTables";
-  public static final String MATERIAL_PSI_ICONS = "PsiIcons";
-  public static final String MATERIAL_FILE_ICONS = "FileIcons";
   public static final String TAB_HIGHLIGHT_POSITION = "TabHighlightPosition";
   @NonNls
   private static final String MIXPANEL_KEY = "mixpanelKey";