+++ src/main/java/com/chrisrm/ideaddlc/MTTitleBarComponent.java
@@ -50,8 +50,6 @@ public final class MTTitleBarComponent implements ProjectComponent {
         setDarkTitleBar();
       }
     });
-
-    connect.subscribe(CustomConfigNotifier.CONFIG_TOPIC, mtCustomThemeConfig -> setDarkTitleBar());
   }
 
   @Override