+++ src/main/java/io/acari/DDLC/DDLCProjectInitializationComponent.kt
@@ -13,42 +13,42 @@ import java.util.*
 /**
  * Forged in the flames of battle by alex.
  */
-class DDLCProjectInitializationComponent(project: Project?) : AbstractProjectComponent(project) {
-    val random = Random(Instant.now().epochSecond / 100)
-    private val mtAddFileColorsAction = DDLCAddFileColorsAction()
+class DDLCProjectInitializationComponent(project: Project) : AbstractProjectComponent(project) {
+  val random = Random(Instant.now().epochSecond / 100)
+  private val mtAddFileColorsAction = DDLCAddFileColorsAction()
 
-    override fun getComponentName(): String {
-        return "DDLCProjectInitializationComponent"
-    }
+  override fun getComponentName(): String {
+    return "DDLCProjectInitializationComponent"
+  }
 
-    override fun projectOpened() {
-        if(MTThemeManager.isDDLCActive()){
-            MTThemeManager.activate()
-            mtAddFileColorsAction.setFileScopes(this.myProject)
-            if(!GeneralSettings.getInstance().isShowTipsOnStartup()){
-                val timesTipsChosen = PropertiesComponent.getInstance().getInt(WRITING_TIP_OF_THE_DAY, 0)
-                if (timesTipsChosen < 1) {
-                    GeneralSettings.getInstance().setShowTipsOnStartup(true)
-                    showMonikasWritingTipOfTheDay(timesTipsChosen)
-                } else if (timesTipsChosen < 2 && shouldShowAgain()) {
-                    showMonikasWritingTipOfTheDay(timesTipsChosen)
-                }
-            }
-        } else {
-            //todo: should probably check to se if the material file scopes are installed (when somebody complains)
-            mtAddFileColorsAction.removeFileScopes(this.myProject)
+  override fun projectOpened() {
+    if (MTThemeManager.isDDLCActive()) {
+      MTThemeManager.activate()
+      mtAddFileColorsAction.setFileScopes(this.myProject)
+      if (!GeneralSettings.getInstance().isShowTipsOnStartup) {
+        val timesTipsChosen = PropertiesComponent.getInstance().getInt(WRITING_TIP_OF_THE_DAY, 0)
+        if (timesTipsChosen < 1) {
+          GeneralSettings.getInstance().isShowTipsOnStartup = true
+          showMonikasWritingTipOfTheDay(timesTipsChosen)
+        } else if (timesTipsChosen < 2 && shouldShowAgain()) {
+          showMonikasWritingTipOfTheDay(timesTipsChosen)
         }
-
-   }
-
-    private fun shouldShowAgain() = random.nextLong() % 42 == 0L
-
-    fun showMonikasWritingTipOfTheDay(timesTipsChosen: Int) {
-        TipDialog.createForProject(this.myProject).show()
-        PropertiesComponent.getInstance().setValue(WRITING_TIP_OF_THE_DAY, (timesTipsChosen + 1).toString())
+      }
+    } else {
+      //todo: should probably check to se if the material file scopes are installed (when somebody complains)
+      mtAddFileColorsAction.removeFileScopes(this.myProject)
     }
+
+  }
+
+  private fun shouldShowAgain() = random.nextLong() % 42 == 0L
+
+  fun showMonikasWritingTipOfTheDay(timesTipsChosen: Int) {
+    TipDialog.createForProject(this.myProject).show()
+    PropertiesComponent.getInstance().setValue(WRITING_TIP_OF_THE_DAY, (timesTipsChosen + 1).toString())
+  }
 
 }
 
 
-val WRITING_TIP_OF_THE_DAY = "WRITING_TIP_OF_THE_DAY"
+const val WRITING_TIP_OF_THE_DAY = "WRITING_TIP_OF_THE_DAY"