+++ src/main/java/io/acari/DDLC/integrations/GithubActions.kt
@@ -1,0 +1,18 @@
+package io.acari.DDLC.integrations
+
+import com.intellij.ide.BrowserUtil
+import java.awt.event.ActionListener
+import java.net.URI
+
+object GithubActions {
+
+  fun createChangelogAction(): ActionListener = ActionListener{
+    val uri = URI("https://github.com/cyclic-reference/ddlc-jetbrains-theme/blob/master/docs/CHANGELOG.md")
+    BrowserUtil.browse(uri)
+  }
+
+  fun createViewIssuesAction(): ActionListener = ActionListener{
+    val uri = URI("https://github.com/cyclic-reference/ddlc-jetbrains-theme/issues")
+    BrowserUtil.browse(uri)
+  }
+}